+++ src/app/api/admin/schedule/start.ts
@@ -1,9 +1,10 @@
 import { Request, Response } from 'express'
-import { respond, schedule } from '~/lib/util'
+import { respond, job } from '~/lib/util'
 
-export async function start(req: Request, res: Response): Promise<void> {
-	// Start schedule
-	schedule.scheduler().start()
+export async function start(req: Request, res: Response): Promise<void>
+{
+	// Start job service
+	await job.start()
 
 	// Success
 	respond(req, res).success()