+++ src/app/api/admin/schedule/stop.ts
@@ -1,9 +1,10 @@
 import { Request, Response } from 'express'
-import { respond, schedule } from '~/lib/util'
+import { respond, job } from '~/lib/util'
 
-export async function stop(req: Request, res: Response): Promise<void> {
-	// Stop schedule
-	schedule.scheduler().stop()
+export async function stop(req: Request, res: Response): Promise<void>
+{
+	// Stop job service
+	await job.close()
 
 	// Success
 	respond(req, res).success()