+++ src/config/index.ts
@@ -4,11 +4,11 @@ import { cache } from './cache'
 import type { CacheConnectionOptions } from '~/lib/util/cache'
 import { db, DbConfig } from './database'
 import { docs, DocsConfig } from './docs'
+import { jobConfig, JobConfig } from './job'
 import { log, LogConfig } from './log'
 import { mail, MailConfig } from './mail'
 import { oauth, OAuthConfig } from './oauth'
 import { search, SearchConfig } from './search'
-import { schedule, ScheduleConfig } from './schedule'
 import { session, SessionConfig } from './session'
 import { socket, SocketConfig } from './socket'
 import { totp, TotpConfig } from './totp'
@@ -20,12 +20,12 @@ export const config: ConfigDictionary = {
 	cache,
 	db,
 	docs,
+	job: jobConfig,
 	log,
 	mail,
 	oauth,
 	search,
 	session,
-	schedule,
 	socket,
 	totp,
 	webhook,
@@ -37,12 +37,12 @@ interface ConfigDictionary {
 	cache: CacheConnectionOptions
 	db: DbConfig
 	docs: DocsConfig
+	job: JobConfig
 	log: LogConfig
 	mail: MailConfig
 	oauth: OAuthConfig
 	search: SearchConfig
 	session: SessionConfig
-	schedule: ScheduleConfig
 	socket: SocketConfig
 	totp: TotpConfig
 	webhook: WebhookConfig