+++ src/lib/util/job/jobs/index.ts
@@ -1,0 +1,1 @@
+export { logMemoryUsage } from './log-memory-usage'