+++ src/lib/util/schedule/types/JobScheduleConfig.ts
@@ -1,6 +1,0 @@
-import { AbstractTask, AbstractTaskGroup } from '../'
-
-export interface JobScheduleConfig {
-	execute: AbstractTask | AbstractTaskGroup
-	schedule?: string
-}